# api/docs/swagger.yaml
basePath: /
definitions:
  apperror.ErrorCode:
    enum:
    - 400
    - 401
    - 404
    - 500
    type: integer
    x-enum-varnames:
    - ErrCodeBadRequest
    - ErrCodeUnauthorized
    - ErrCodeNotFound
    - ErrCodeInternal
  apperror.Response:
    properties:
      message:
        type: string
      status_code:
        $ref: '#/definitions/apperror.ErrorCode'
    type: object
...
  dto.AuthInfo:
    properties:
      fee:
        $ref: '#/definitions/dto.Fee'
      signer_infos:
        items:
          $ref: '#/definitions/dto.SignerInfo'
        type: array
    type: object
...
  dto.Body:
    properties:
      extension_options:
        items: {}
        type: array
      memo:
        type: string
      messages:
        items:
          additionalProperties: {}
          type: object
        type: array
      non_critical_extension_options:
        items: {}
        type: array
      timeout_height:
        type: string
    type: object
  dto.Coin:
    properties:
      amount:
        type: string
      denom:
        type: string
    type: object
...
  dto.Event:
    properties:
      attributes:
        items:
          $ref: '#/definitions/dto.EventAttribute'
        type: array
      type:
        type: string
    type: object
  dto.EventAttribute:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  dto.Fee:
    properties:
      amount:
        items:
          $ref: '#/definitions/dto.Coin'
        type: array
      gas_limit:
        type: string
      granter:
        type: string
      payer:
        type: string
    type: object
  dto.Log:
    properties:
      events:
        items:
          $ref: '#/definitions/dto.Event'
        type: array
      log:
        description: Can be string or map[string]string
      msg_index:
        type: integer
    type: object
...
  dto.PaginationResponse:
    properties:
      next_key:
        type: string
      previous_key:
        type: string
      total:
        type: string
    type: object
...
  dto.ProposalAnswerCounts:
    properties:
      abstain:
        type: integer
      "no":
        type: integer
      no_with_veto:
        type: integer
      total:
        type: integer
      weighted:
        type: integer
      "yes":
        type: integer
    type: object
  dto.ProposalAnswerCountsResponse:
    properties:
      all:
        $ref: '#/definitions/dto.ProposalAnswerCounts'
      validator:
        $ref: '#/definitions/dto.ProposalValidatorAnswerCounts'
    type: object
  dto.ProposalContent:
    properties:
      messages:
        type: object
      metadata:
        type: string
    type: object
  dto.ProposalDeposit:
    properties:
      amount:
        items:
          $ref: '#/definitions/dto.Coin'
        type: array
      depositor:
        type: string
      timestamp:
        type: string
      tx_hash:
        type: string
    type: object
  dto.ProposalInfo:
    properties:
      abstain:
        type: string
      content:
        $ref: '#/definitions/dto.ProposalContent'
      created_height:
        type: integer
      created_timestamp:
        type: string
      created_tx_hash:
        type: string
      deposit_end_time:
        type: string
      description:
        type: string
      failed_reason:
        type: string
      id:
        type: integer
      is_emergency:
        type: boolean
      is_expedited:
        type: boolean
      messages:
        items:
          type: integer
        type: array
      metadata:
        type: string
      "no":
        type: string
      no_with_veto:
        type: string
      proposal_deposits:
        items:
          $ref: '#/definitions/dto.ProposalDeposit'
        type: array
      proposer:
        type: string
      resolved_height:
        type: integer
      resolved_timestamp:
        type: string
      resolved_total_voting_power:
        type: string
      status:
        type: string
      submit_time:
        type: string
      title:
        type: string
      total_deposit:
        items:
          $ref: '#/definitions/dto.Coin'
        type: array
      types:
        items:
          type: integer
        type: array
      version:
        type: string
      voting_end_time:
        type: string
      voting_time:
        type: string
      "yes":
        type: string
    type: object
  dto.ProposalInfoResponse:
    properties:
      info:
        $ref: '#/definitions/dto.ProposalInfo'
    type: object
  dto.ProposalSummary:
    properties:
      deposit_end_time:
        type: string
      id:
        type: integer
      is_emergency:
        type: boolean
      is_expedited:
        type: boolean
      proposer:
        type: string
      resolved_height:
        type: integer
      status:
        type: string
      title:
        type: string
      types:
        type: object
      voting_end_time:
        type: string
    type: object
  dto.ProposalValidatorAnswerCounts:
    properties:
      abstain:
        type: integer
      did_not_vote:
        type: integer
      "no":
        type: integer
      no_with_veto:
        type: integer
      total:
        type: integer
      total_validators:
        type: integer
      weighted:
        type: integer
      "yes":
        type: integer
    type: object
  dto.ProposalValidatorVote:
    properties:
      identity:
        type: string
      moniker:
        type: string
      validator_address:
        type: string
    type: object
  dto.ProposalValidatorVotesResponse:
    properties:
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
      votes:
        items:
          $ref: '#/definitions/dto.ProposalVote'
        type: array
    type: object
  dto.ProposalVote:
    properties:
      abstain:
        type: number
      is_vote_weighted:
        type: boolean
      "no":
        type: number
      no_with_veto:
        type: number
      proposal_id:
        type: integer
      timestamp:
        type: string
      tx_hash:
        type: string
      validator:
        $ref: '#/definitions/dto.ProposalValidatorVote'
      voter:
        type: string
      "yes":
        type: number
    type: object
  dto.ProposalVotesResponse:
    properties:
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
      votes:
        items:
          $ref: '#/definitions/dto.ProposalVote'
        type: array
    type: object
  dto.ProposalsResponse:
    properties:
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
      proposals:
        items:
          $ref: '#/definitions/dto.ProposalSummary'
        type: array
    type: object
  dto.PublicKey:
    properties:
      '@type':
        type: string
      key:
        type: string
    type: object
  dto.SignerInfo:
    properties:
      mode_info: {}
      public_key:
        $ref: '#/definitions/dto.PublicKey'
      sequence:
        type: string
    type: object
  dto.Tx:
    properties:
      auth_info:
        $ref: '#/definitions/dto.AuthInfo'
      body:
        $ref: '#/definitions/dto.Body'
      signatures:
        items:
          type: string
        type: array
    type: object
  dto.TxByHashResponse:
    properties:
      tx:
        $ref: '#/definitions/dto.Tx'
      tx_response:
        $ref: '#/definitions/dto.TxResponse'
    type: object
  dto.TxModel:
    properties:
      hash:
        type: string
      height:
        type: integer
      is_ibc:
        type: boolean
      is_opinit:
        type: boolean
      is_send:
        type: boolean
      messages:
        type: object
      sender:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  dto.TxResponse:
    properties:
      code:
        type: integer
      codespace:
        type: string
      data:
        type: string
      events:
        items:
          $ref: '#/definitions/dto.Event'
        type: array
      gas_used:
        type: string
      gas_wanted:
        type: string
      height:
        type: string
      info:
        type: string
      logs:
        items:
          $ref: '#/definitions/dto.Log'
        type: array
      raw_log:
        type: string
      timestamp:
        description: unix time (GMT)
        type: string
      tx:
        $ref: '#/definitions/dto.Tx'
      txhash:
        type: string
    type: object
  dto.TxsModelResponse:
    properties:
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
      txs:
        items:
          $ref: '#/definitions/dto.TxModel'
        type: array
    type: object
  dto.TxsResponse:
    properties:
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
      txs:
        items:
          $ref: '#/definitions/dto.TxResponse'
        type: array
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API service for the Core Indexer project
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Core Indexer API
  version: "1.0"
paths:
...
  /indexer/proposal/v1/proposals:
    get:
      description: Retrieve the list of all proposals
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: pagination.offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: pagination.limit
        type: integer
      - default: false
        description: Reverse order for pagination
        in: query
        name: pagination.reverse
        type: boolean
      - default: false
        description: Count total number of transactions
        in: query
        name: pagination.count_total
        type: boolean
      - default: ""
        description: Filter proposals by proposer
        in: query
        name: proposer
        type: string
      - default: ""
        description: Filter proposals by status(es)
        in: query
        name: statuses
        type: string
      - default: ""
        description: Filter proposals by proposal type
        in: query
        name: types
        type: string
      - default: ""
        description: Search proposals by title or exact proposal id
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProposalsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get list of proposals
      tags:
      - Proposal
  /indexer/proposal/v1/proposals/{proposalId}/answer_counts:
    get:
      description: Retrieve vote counts of a proposal by options
      parameters:
      - description: Proposal Id
        in: path
        name: proposalId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProposalAnswerCountsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get votes count of a proposal
      tags:
      - Proposal
  /indexer/proposal/v1/proposals/{proposalId}/info:
    get:
      description: Retrieve proposal details
      parameters:
      - description: Proposal Id
        in: path
        name: proposalId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProposalInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get proposal info
      tags:
      - Proposal
  /indexer/proposal/v1/proposals/{proposalId}/validator_votes:
    get:
      description: Retrieve list of all proposal votes by validators
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: pagination.offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: pagination.limit
        type: integer
      - default: false
        description: Reverse order for pagination
        in: query
        name: pagination.reverse
        type: boolean
      - default: false
        description: Count total number of transactions
        in: query
        name: pagination.count_total
        type: boolean
      - description: Proposal Id
        in: path
        name: proposalId
        required: true
        type: string
      - default: ""
        description: Search proposal vote by validator moniker or address
        in: query
        name: search
        type: string
      - default: ""
        description: Search proposal votes by vote option
        in: query
        name: answer
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProposalValidatorVotesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get validator votes of a proposal
      tags:
      - Proposal
  /indexer/proposal/v1/proposals/{proposalId}/votes:
    get:
      description: Retrieve list of all proposal votes
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: pagination.offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: pagination.limit
        type: integer
      - default: false
        description: Reverse order for pagination
        in: query
        name: pagination.reverse
        type: boolean
      - default: false
        description: Count total number of transactions
        in: query
        name: pagination.count_total
        type: boolean
      - description: Proposal Id
        in: path
        name: proposalId
        required: true
        type: string
      - default: ""
        description: Search proposal vote by voter address
        in: query
        name: search
        type: string
      - default: ""
        description: Search proposal votes by vote option
        in: query
        name: answer
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProposalVotesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get all proposal votes
      tags:
      - Proposal
  /indexer/proposal/v1/proposals/types:
    get:
      description: Retrieve all submitted proposal types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get list of submitted proposal types
      tags:
      - Proposal
  /indexer/tx/v1/txs:
    get:
      consumes:
      - application/json
      description: Retrieve a list of transactions with pagination
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: pagination.offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: pagination.limit
        type: integer
      - default: false
        description: Reverse order for pagination
        in: query
        name: pagination.reverse
        type: boolean
      - default: false
        description: Count total number of transactions
        in: query
        name: pagination.count_total
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TxsModelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get transactions
      tags:
      - Transaction
  /indexer/tx/v1/txs/{tx_hash}:
    get:
      consumes:
      - application/json
      description: Retrieve transaction details by its hash
      parameters:
      - description: Transaction hash
        in: path
        name: tx_hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TxByHashResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get transaction by hash
      tags:
      - Transaction
  /indexer/tx/v1/txs/by_account/{accountAddress}:
    get:
      description: Retrieve transactions by account address
      parameters:
      - description: Account address
        in: path
        name: accountAddress
        required: true
        type: string
      - default: 0
        description: Offset for pagination
        in: query
        name: pagination.offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: pagination.limit
        type: integer
      - default: false
        description: Reverse order for pagination
        in: query
        name: pagination.reverse
        type: boolean
      - default: false
        description: Count total number of transactions
        in: query
        name: pagination.count_total
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TxsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get transactions by account address
      tags:
      - Transaction
  /indexer/tx/v1/txs/count:
    get:
      description: Retrieve the total number of transactions
      produces:
      - application/json
      responses:
        "200":
          description: Transaction count
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.Response'
      summary: Get transaction count
      tags:
      - Transaction
...
swagger: "2.0"
...